+++ client/eslint.config.js
@@ -4,6 +4,8 @@ import reactHooks from 'eslint-plugin-react-hooks'
 import reactRefresh from 'eslint-plugin-react-refresh'
 import { defineConfig, globalIgnores } from 'eslint/config'
 
+import i18next from 'eslint-plugin-i18next'
+
 export default defineConfig([
   globalIgnores(['dist']),
   {
@@ -12,6 +14,7 @@ export default defineConfig([
       js.configs.recommended,
       reactHooks.configs.flat.recommended,
       reactRefresh.configs.vite,
+      i18next.configs.recommended,
     ],
     languageOptions: {
       ecmaVersion: 2020,
@@ -24,6 +27,7 @@ export default defineConfig([
     },
     rules: {
       'no-unused-vars': ['error', { varsIgnorePattern: '^[A-Z_]' }],
+      'i18next/no-literal-string': ['warn', { markupOnly: true }],
     },
   },
 ])